refactor(@formatjs/intl-numberformat): use generated Test262 harness rules
