+++ packages/intl-datetimeformat/test262-baseline.json
@@ -87,8 +87,8 @@
     "intl402/DateTimeFormat/prototype/format/timedatestyle-en.js (strict mode)": "Result for full with {} Expected SameValue(«\"2:12:47 PM Coordinated Universal Time\"», «\"14:12:47 Coordinated Universal Time\"») to be true",
     "intl402/DateTimeFormat/prototype/formatRange/argument-near-time-boundaries.js (default)": "Expected a RangeError but got a Error",
     "intl402/DateTimeFormat/prototype/formatRange/argument-near-time-boundaries.js (strict mode)": "Expected a RangeError but got a Error",
-    "intl402/DateTimeFormat/prototype/formatRange/argument-to-integer.js (default)": "formatRange(-0.9) Expected SameValue(«\"<time> م – <time> ص\"», «\"<time> ص – <time> ص\"») to be true",
-    "intl402/DateTimeFormat/prototype/formatRange/argument-to-integer.js (strict mode)": "formatRange(-0.9) Expected SameValue(«\"<time> م – <time> ص\"», «\"<time> ص – <time> ص\"») to be true",
+    "intl402/DateTimeFormat/prototype/formatRange/argument-to-integer.js (default)": "formatRange(-0.9) Expected SameValue(«\"11:59:59 م – <current endpoint>\"», «\"12:00:00 ص – <current endpoint>\"») to be true",
+    "intl402/DateTimeFormat/prototype/formatRange/argument-to-integer.js (strict mode)": "formatRange(-0.9) Expected SameValue(«\"11:59:59 م – <current endpoint>\"», «\"12:00:00 ص – <current endpoint>\"») to be true",
     "intl402/DateTimeFormat/prototype/formatRange/builtin.js (default)": "Test262Error {\n  message: \"Built-in functions that aren't constructors must not have a prototype property. Expected SameValue(«true», «false») to be true\"\n}",
     "intl402/DateTimeFormat/prototype/formatRange/builtin.js (strict mode)": "Test262Error {\n  message: \"Built-in functions that aren't constructors must not have a prototype property. Expected SameValue(«true», «false») to be true\"\n}",
     "intl402/DateTimeFormat/prototype/formatRange/date-same-returns-single-date.js (default)": "same output with time options Expected SameValue(«\"1:02:03 AM – 1:02:03 AM\"», «\"1:02:03 AM\"») to be true",
