+++ tools/test262/defs.bzl
@@ -1,25 +1,61 @@
-"""Test262 gates report tracked failures separately from passing tests."""
+"""Generated harness execution with separate baseline validation."""
 
-load("@aspect_rules_js//js:defs.bzl", "js_test")
+load("@aspect_rules_js//js:defs.bzl", "js_run_binary", "js_test")
+load("@npm//:test262-harness/package_json.bzl", test262_harness_bin = "bin")
 
 def test262_test(name, suite, prelude, data, baseline = "test262-baseline.json"):
-    """Run complete suites, with separate strict and native control targets."""
-    for mode in ["baseline", "strict", "native"]:
-        args = ["--suite", suite, "--root", "../+http_archive+com_github_tc39_test262"]
-        inputs = data + ["//:node_modules/minimist", "//:node_modules/test262-harness"]
-        if mode != "native":
-            args += ["--prelude", "$(rootpath %s)" % prelude]
-            inputs += [prelude]
-        if mode == "baseline":
-            args += ["--baseline", "$(rootpath %s)" % baseline]
-            inputs += [baseline]
-        else:
-            args += ["--" + mode]
-        js_test(
-            name = name + ("" if mode == "baseline" else "-" + mode),
-            entry_point = "//tools/test262:runner",
-            args = args,
-            data = inputs,
+    """Keep direct strict/native tests and validate captured baseline reports."""
+    root = "../+http_archive+com_github_tc39_test262"
+    realm_prelude = name + "-realm-prelude.js"
+    js_run_binary(
+        name = name + "-realm-prelude",
+        tool = "//tools/test262:prelude",
+        srcs = [prelude],
+        outs = [realm_prelude],
+        args = ["--input", "$(rootpath %s)" % prelude, "--out", "$(rootpath %s)" % realm_prelude],
+    )
+    args = [
+        "--reporter",
+        "json",
+        "--reporter-keys",
+        "file,scenario,result",
+        "--errorForFailures",
+        "--timeout",
+        "30000",
+        "--test262Dir",
+        root,
+        root + "/test/" + suite + "/**/*.js",
+    ]
+    polyfill_args = args + ["--prelude", "$(rootpath %s)" % realm_prelude]
+    report = name + "-results.json"
+    status = name + "-exit-code.txt"
+    stderr = name + "-stderr.txt"
+
+    # Capture the real failing exit status; only the validation test decides
+    # whether those failures match the reviewed baseline.
+    test262_harness_bin.test262_harness(
+        name = name + "-report",
+        srcs = data + [realm_prelude],
+        # Build actions resolve external inputs below bazel-bin/external;
+        # test runfiles resolve them beside the main repository.
+        args = [arg.replace(root, "external/" + root[3:]) for arg in polyfill_args],
+        stdout = report,
+        stderr = stderr,
+        exit_code_out = status,
+        env = {"TZ": "UTC"},
+    )
+    js_test(
+        name = name,
+        entry_point = "//tools/test262:validate",
+        data = [report, status, stderr, baseline, "//tools/test262:validation_sources", "//:node_modules/minimist"],
+        args = ["--suite", suite, "--report", "$(rootpath %s)" % report, "--status", "$(rootpath %s)" % status, "--stderr", "$(rootpath %s)" % stderr, "--baseline", "$(rootpath %s)" % baseline],
+    )
+    for mode in ["strict", "native"]:
+        test262_harness_bin.test262_harness_test(
+            name = name + "-" + mode,
+            args = polyfill_args if mode == "strict" else args,
+            data = data + ([realm_prelude] if mode == "strict" else []),
+            env = {"TZ": "UTC"},
             size = "large",
-            tags = [] if mode == "baseline" else ["manual"],
+            tags = ["manual"],
         )
