+++ tools/test262/fixtures/harness/assert.js
@@ -1,0 +1,1 @@
+// Minimal harness fixture.
