+++ tools/test262/fixtures/package.json
@@ -1,0 +1,3 @@
+{
+  "version": "5.0.0"
+}
