+++ tools/test262/fixtures/prelude.js
@@ -1,0 +1,1 @@
+globalThis.polyfillMarker = 42
