+++ tools/test262/fixtures/test/fail.js
@@ -1,0 +1,4 @@
+/*---
+description: generated harness fixture
+---*/
+throw new Error('intentional')
