+++ tools/test262/fixtures/test/pass.js
@@ -1,0 +1,4 @@
+/*---
+description: generated harness fixture
+---*/
+void 0
